Program/requirements.txt
# Requirements for Production Tracking Data System (D-CONDA)
# Install with: pip install -r requirements.txt

Pillow==12.1.1
openpyxl==3.1.5
eel==0.18.2
